pyproject.toml
[tool.poetry]
name = "groot"
version = "1.0.0"
description = "A Discord bot that responds when you say Groot's name."
authors = ["jonxola"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
